final firebase & database changer
final firebase & database changer

+++ app/src/main/java/com/example/baytna/Model/WorkerItemsHome.kt
@@ -3,9 +3,8 @@ package com.example.baytna.Model
 data class WorkerItemsHome (
     var Name:String ="",
     var ImagePath:String = "",
-    var Rate:Long =0,
+    var Rate:Double =0.0,
     val CategoryId:Long = 0,
     var price:String = "",
     var Mobile :String = "",
-    var Id : Long = 0
-)
+    var Id : Long = 0)
